Bugfix: avoid flooding lots of 'OK' statuses at start.
When we first start, we won't know the status for each service yet, so don't
spam a buttload of 'OK' messages to the channel, as that will just piss people
off, and make them want to stab me and/or the bot administrator in the face with
a fork.  Nobody wants a fork to the face.  (Rusty spoon, maybe, but not a fork.)