Support units for durations; refactor validation.
Refactor validation to use a coderef always, which returns a possibly-modified
version of the value if it was acceptable, or undef if it was no good.

This lets the validator for durations accept e.g. 10m and return 600, so
durations can be supplied in a human-friendly way, but turned into seconds
automatically.