Report update failures.
If report_statuses includes UPDATE_FAIL, then announce failures to fetch/scrape
the Nagios status page, rather than staying silent and making it seem that all
is fine when potentially it isn't.

TODO: maybe add some robustification, e.g. auto-retry and only announce the
failure if a couple of attempts in a row failed, to avoid potential for spurious
noise.