Factor Nagios-polling out of tick()
I want to be able to add a "check now" command to force an update.

I also want to further factor out the "get a list of hosts" stuff into a method
that could be called when adding a new instance to monitor, so we can refuse to
add it if we can't talk to it with the given details.